@@ -1,15 +1,16 @@
 {
   "name": "vuero",
-  "version": "2.8.1",
+  "version": "2.9.0",
   "private": true,
   "type": "module",
   "license": "SEE LICENSE IN LICENSE.md",
+  "packageManager": "pnpm@8.14.0",
   "author": {
     "name": "cssninjaStudio <hello@cssninja.io> (https://cssninja.io)"
   },
   "engines": {
-    "node": ">=16.15.0",
-    "npm": ">=8"
+    "node": ">=18",
+    "npm": ">=10"
   },
   "gitHooks": {
     "pre-commit": "lint-staged",
@@ -71,161 +63,162 @@
     "cypress": "cypress"
   },
   "dependencies": {
-    "@ckeditor/ckeditor5-build-classic": "39.0.1",
+    "@ckeditor/ckeditor5-build-classic": "41.0.0",
     "@ckeditor/ckeditor5-vue": "5.1.0",
     "@cssninja/bulma-css-vars": "0.9.2",
     "@iconify/iconify": "3.1.1",
-    "@mapbox/mapbox-gl-geocoder": "5.0.1",
-    "@minify-html/node": "^0.11.1",
+    "@mapbox/mapbox-gl-geocoder": "5.0.2",
+    "@minify-html/node": "^0.15.0",
+    "@nuxt/devalue": "2.0.2",
     "@popperjs/core": "2.11.8",
-    "@unhead/ssr": "1.3.7",
-    "@unhead/vue": "1.3.7",
-    "@vee-validate/zod": "4.11.3",
-    "@vueform/multiselect": "2.6.2",
-    "@vueform/slider": "2.1.7",
-    "@vueuse/core": "10.4.0",
-    "@vueuse/router": "10.4.0",
-    "apexcharts": "3.42.0",
-    "axios": "1.6.0",
-    "billboard.js": "3.9.3",
-    "bulma": "0.9.4",
-    "datatables.net": "^1.13.6",
-    "datatables.net-bs5": "^1.13.6",
-    "datatables.net-buttons": "^2.4.2",
-    "datatables.net-buttons-dt": "^2.4.2",
-    "datatables.net-colreorder-dt": "^1.7.0",
+    "@unhead/addons": "1.8.10",
+    "@unhead/ssr": "1.8.10",
+    "@unhead/vue": "1.8.10",
+    "@vee-validate/zod": "4.12.4",
+    "@vueform/multiselect": "2.6.6",
+    "@vueform/slider": "2.1.10",
+    "@vueuse/core": "10.7.2",
+    "@vueuse/router": "10.7.2",
+    "apexcharts": "3.45.2",
+    "axios": "^1.6.7",
+    "billboard.js": "3.10.3",
+    "bulma": "npm:@cssninja/bulma@0.9.4",
+    "compression": "1.7.4",
+    "datatables.net": "^2.0.0",
+    "datatables.net-bs5": "^2.0.0",
+    "datatables.net-buttons": "^3.0.0",
+    "datatables.net-colreorder": "^2.0.0",
     "datatables.net-fixedheader-dt": "^3.4.0",
-    "datatables.net-select": "^1.7.0",
+    "datatables.net-select": "^2.0.0",
     "datatables.net-vue3": "^2.1.3",
-    "dayjs": "1.11.9",
-    "defu": "6.1.2",
+    "dayjs": "1.11.10",
+    "defu": "6.1.4",
     "dragula": "3.7.3",
     "dropzone": "6.0.0-beta.2",
-    "filepond": "4.30.4",
+    "filepond": "4.30.6",
     "filepond-plugin-file-validate-size": "2.2.8",
     "filepond-plugin-file-validate-type": "1.2.8",
     "filepond-plugin-image-crop": "2.0.6",
     "filepond-plugin-image-edit": "1.6.3",
     "filepond-plugin-image-exif-orientation": "1.0.11",
-    "filepond-plugin-image-preview": "4.6.11",
+    "filepond-plugin-image-preview": "4.6.12",
     "filepond-plugin-image-resize": "2.0.10",
     "filepond-plugin-image-transform": "3.8.7",
-    "focus-trap": "7.5.2",
-    "focus-trap-vue": "4.0.2",
-    "imask": "7.1.3",
-    "mapbox-gl": "2.15.0",
+    "focus-trap": "7.5.4",
+    "focus-trap-vue": "4.0.3",
+    "h3": "1.10.0",
+    "highlight.js": "^11.9.0",
+    "imask": "7.3.0",
+    "jiti": "1.21.0",
+    "jszip": "^3.10.1",
+    "listhen": "1.5.6",
+    "mapbox-gl": "3.1.0",
     "notyf": "3.10.0",
     "nprogress": "0.2.0",
     "ofetch": "1.3.3",
     "pdfjs-dist": "^3.4.120",
-    "pdfmake": "^0.2.7",
-    "photoswipe": "5.3.8",
-    "pinia": "2.1.6",
-    "simple-datatables": "7.3.0",
+    "pdfmake": "0.3.0-beta.7",
+    "photoswipe": "5.4.3",
+    "pinia": "2.1.7",
+    "plyr": "3.7.8",
+    "rehype-autolink-headings": "7.1.0",
+    "rehype-external-links": "3.0.0",
+    "rehype-raw": "7.0.0",
+    "rehype-sanitize": "6.0.0",
+    "rehype-shikiji": "0.10.0",
+    "rehype-slug": "6.0.0",
+    "rehype-stringify": "10.0.0",
+    "remark-frontmatter": "5.0.0",
+    "remark-gfm": "4.0.0",
+    "remark-parse": "11.0.0",
+    "remark-rehype": "11.1.0",
+    "serve": "14.2.1",
+    "serve-static": "1.15.0",
     "simplebar": "6.2.5",
-    "simplebar-vue": "2.3.3",
     "textarea-markdown-editor": "1.0.5-rc.4",
     "tiny-slider": "2.9.4",
     "tippy.js": "6.3.7",
     "tslib": "2.6.2",
-    "unhead": "1.3.7",
+    "unhead": "1.8.10",
+    "unplugin-vue-router": "0.7.0",
     "v-calendar": "3.0.3",
-    "vee-validate": "4.11.3",
+    "vee-validate": "4.12.4",
     "vivus": "0.4.6",
-    "vue": "3.3.4",
-    "vue-accessible-color-picker": "4.1.4",
-    "vue-i18n": "9.3.0-beta.25",
-    "vue-router": "4.2.4",
+    "vue": "3.4.15",
+    "vue-accessible-color-picker": "5.0.1",
+    "vue-i18n": "9.9.0",
+    "vue-router": "4.2.5",
     "vue-scrollto": "2.20.0",
-    "vue-tippy": "6.3.1",
+    "vue-tippy": "6.4.1",
     "vue3-apexcharts": "1.4.4",
+    "vue3-google-map": "^0.18.0",
     "vue3-pdfjs": "^0.1.6",
     "xlsx": "^0.18.5",
     "zod": "3.22.4"
   },
   "devDependencies": {
-    "@commitlint/cli": "17.7.1",
-    "@commitlint/config-conventional": "17.7.0",
-    "@iconify/json": "2.2.106",
-    "@intlify/unplugin-vue-i18n": "0.12.3",
-    "@nuxt/devalue": "2.0.2",
-    "@stefanprobst/remark-shiki": "2.2.1",
-    "@types/dragula": "3.7.1",
-    "@types/fs-extra": "11.0.1",
-    "@types/js-yaml": "4.0.5",
-    "@types/mapbox-gl": "2.7.13",
-    "@types/mapbox__mapbox-gl-geocoder": "4.7.4",
-    "@types/node": "20.4.5",
-    "@types/nprogress": "0.2.0",
-    "@types/pdfmake": "^0.2.4",
-    "@types/photoswipe": "4.1.2",
-    "@types/prismjs": "1.26.0",
-    "@types/vivus": "0.4.4",
-    "@typescript-eslint/eslint-plugin": "6.4.1",
-    "@typescript-eslint/parser": "6.4.1",
-    "@unhead/addons": "1.3.7",
-    "@vitejs/plugin-vue": "4.3.3",
-    "commitlint": "17.7.1",
-    "compression": "1.7.4",
+    "@ckeditor/ckeditor5-font": "^41.0.0",
+    "@commitlint/cli": "18.5.0",
+    "@commitlint/config-conventional": "18.5.0",
+    "@iconify/json": "2.2.172",
+    "@intlify/unplugin-vue-i18n": "2.0.0",
+    "@stylistic/eslint-plugin": "^1.5.4",
+    "@types/datatables.net-fixedheader": "^3.1.8",
+    "@types/dragula": "3.7.4",
+    "@types/fs-extra": "11.0.4",
+    "@types/html-minifier-terser": "^7.0.2",
+    "@types/js-yaml": "4.0.9",
+    "@types/mapbox-gl": "2.7.19",
+    "@types/mapbox__mapbox-gl-geocoder": "4.7.7",
+    "@types/node": "20.11.5",
+    "@types/nprogress": "0.2.3",
+    "@types/pdfmake": "^0.2.9",
+    "@types/photoswipe": "4.1.6",
+    "@types/vivus": "0.4.7",
+    "@typescript-eslint/eslint-plugin": "6.19.1",
+    "@typescript-eslint/parser": "6.19.1",
+    "@vitejs/plugin-vue": "5.0.3",
+    "commitlint": "18.5.0",
     "cross-env": "7.0.3",
-    "cypress": "12.17.4",
-    "eslint": "8.48.0",
-    "eslint-config-prettier": "9.0.0",
+    "cypress": "13.6.3",
+    "eslint": "8.56.0",
     "eslint-plugin-frontmatter": "0.0.8",
     "eslint-plugin-md": "1.0.19",
-    "eslint-plugin-prettier": "5.0.0",
-    "eslint-plugin-vue": "9.17.0",
-    "eslint-plugin-vuejs-accessibility": "2.2.0",
-    "fs-extra": "11.1.1",
+    "eslint-plugin-sonarjs": "^0.23.0",
+    "eslint-plugin-vue": "9.20.1",
+    "eslint-plugin-vuejs-accessibility": "2.2.1",
+    "fs-extra": "11.2.0",
     "gray-matter": "4.0.3",
-    "h3": "1.8.0",
-    "jiti": "1.19.3",
+    "html-minifier-terser": "^7.2.0",
     "js-yaml": "4.1.0",
-    "json-server": "0.17.3",
-    "lint-staged": "14.0.1",
-    "listhen": "1.4.0",
-    "magic-string": "0.30.3",
+    "json-server": "0.17.4",
+    "lint-staged": "15.2.0",
+    "magic-string": "0.30.5",
     "npm-run-all": "4.1.5",
-    "plyr": "3.7.8",
-    "postcss-html": "1.5.0",
-    "postcss-scss": "4.0.7",
-    "prettier": "3.0.2",
-    "prismjs": "1.29.0",
-    "rehype-autolink-headings": "6.1.1",
-    "rehype-external-links": "2.1.0",
-    "rehype-raw": "7.0.0",
-    "rehype-sanitize": "6.0.0",
-    "rehype-slug": "5.1.0",
-    "rehype-stringify": "10.0.0",
-    "remark-frontmatter": "4.0.1",
-    "remark-gfm": "3.0.1",
-    "remark-parse": "10.0.2",
-    "remark-rehype": "10.1.0",
-    "rimraf": "5.0.1",
+    "postcss-html": "1.6.0",
+    "postcss-scss": "4.0.9",
+    "rimraf": "5.0.5",
     "rollup-plugin-purgecss": "5.0.0",
-    "sass": "1.62.0",
-    "scule": "1.0.0",
-    "serve": "14.2.1",
-    "serve-static": "1.15.0",
-    "shiki": "0.14.3",
-    "stylelint": "15.10.3",
+    "sass": "npm:sass-embedded@1.69.5",
+    "scule": "1.2.0",
+    "shikiji": "0.10.0",
+    "stylelint": "16.2.0",
     "stylelint-config-recommended-vue": "1.5.0",
-    "stylelint-config-standard": "34.0.0",
-    "stylelint-config-standard-scss": "10.0.0",
-    "stylelint-prettier": "4.0.2",
-    "stylelint-scss": "5.1.0",
-    "typescript": "5.2.2",
-    "unified": "11.0.2",
-    "unplugin-auto-import": "0.16.6",
-    "unplugin-fonts": "1.0.3",
-    "unplugin-vue-components": "0.25.1",
-    "unplugin-vue-router": "0.6.4",
-    "vite": "5.0.10",
+    "stylelint-config-standard": "36.0.0",
+    "stylelint-config-standard-scss": "13.0.0",
+    "stylelint-scss": "6.1.0",
+    "typescript": "5.3.3",
+    "unified": "11.0.4",
+    "unplugin-auto-import": "0.17.3",
+    "unplugin-fonts": "1.1.1",
+    "unplugin-vue-components": "0.26.0",
+    "vite": "5.0.12",
     "vite-plugin-imagemin": "0.6.1",
-    "vite-plugin-purge-icons": "0.9.2",
-    "vite-plugin-pwa": "0.16.4",
-    "vite-plugin-radar": "0.9.1",
-    "vue-tsc": "1.8.8",
+    "vite-plugin-purge-icons": "0.10.0",
+    "vite-plugin-pwa": "0.17.4",
+    "vite-plugin-radar": "0.9.2",
+    "vue-component-meta": "^1.8.27",
+    "vue-tsc": "1.8.27",
     "yorkie": "2.0.0"
   },
   "pnpm": {
