@@ -4,7 +4,7 @@ import { useUserSession } from '/@src/stores/userSession'
 
 export function useFetch() {
   return ofetch.create({
-    baseURL: import.meta.env.VITE_API_BASE_URL ?? 'http://localhost:8080',
+    baseURL: import.meta.env.VITE_API_BASE_URL + '/api/v1/' ?? 'http://localhost:8080',
     // We set an interceptor for each request to
     // include Bearer token to the request if user is logged in
     onRequest: ({ options }) => {
@@ -13,6 +13,8 @@ export function useFetch() {
       if (userSession.isLoggedIn) {
         options.headers = {
           ...options.headers,
+          'Content-Type': 'application/json',
+          'X-Requested-With': 'XMLHttpRequest',
           Authorization: `Bearer ${userSession.token}`,
         }
       }
