@@ -14,7 +14,7 @@ declare module 'vue' {
     DashboardPrimary: typeof import('./../src/components/pages/dashboards/DashboardPrimary.vue')['default']
     DashboardsMobileSubsidebar: typeof import('./../src/layouts/mobile-subsidebars/DashboardsMobileSubsidebar.vue')['default']
     DashboardsSubsidebar: typeof import('./../src/layouts/sidebar-subsidebar/DashboardsSubsidebar.vue')['default']
-    DataTableWrapper: typeof import('./../src/components/base/data-table/DataTableWrapper.vue')['default']
+    DataTableWrapper: typeof import('./../src/components/base/DataTableWrapper.vue')['default']
     DepartmentList: typeof import('./../src/components/pages/tool/departments/DepartmentList.vue')['default']
     IcoLogo: typeof import('./../src/components/svg/IcoLogo.vue')['default']
     LandingEmptyNavigation: typeof import('./../src/components/partials/landing/LandingEmptyNavigation.vue')['default']
@@ -41,80 +41,82 @@ declare module 'vue' {
     UIWidget: typeof import('./../src/components/partials/widgets/ui-creative/UIWidget.vue')['default']
     UserProfileDropdown: typeof import('./../src/components/partials/dropdowns/UserProfileDropdown.vue')['default']
     UsersList: typeof import('./../src/components/pages/setting/users/UsersList.vue')['default']
-    VAccordion: typeof import('./../src/components/base/accordion/VAccordion.vue')['default']
-    VAccordionImage: typeof import('./../src/components/base/accordion/VAccordionImage.vue')['default']
-    VAction: typeof import('./../src/components/base/button/VAction.vue')['default']
-    VAnimatedCheckbox: typeof import('./../src/components/base/form/VAnimatedCheckbox.vue')['default']
-    VAvatar: typeof import('./../src/components/base/avatar/VAvatar.vue')['default']
-    VAvatarStack: typeof import('./../src/components/base/avatar/VAvatarStack.vue')['default']
-    VBillboardJS: typeof import('./../src/components/base/plugins/VBillboardJS.vue')['default']
-    VBlock: typeof import('./../src/components/base/block/VBlock.vue')['default']
-    VBreadcrumb: typeof import('./../src/components/base/breadcrumb/VBreadcrumb.vue')['default']
-    VButton: typeof import('./../src/components/base/button/VButton.vue')['default']
-    VButtons: typeof import('./../src/components/base/button/VButtons.vue')['default']
-    VCard: typeof import('./../src/components/base/card/VCard.vue')['default']
-    VCardAction: typeof import('./../src/components/base/card/VCardAction.vue')['default']
-    VCardAdvanced: typeof import('./../src/components/base/card/VCardAdvanced.vue')['default']
-    VCardMedia: typeof import('./../src/components/base/card/VCardMedia.vue')['default']
-    VCardSocial: typeof import('./../src/components/base/card/VCardSocial.vue')['default']
-    VCheckbox: typeof import('./../src/components/base/form/VCheckbox.vue')['default']
-    VCollapse: typeof import('./../src/components/base/accordion/VCollapse.vue')['default']
-    VCollapseLinks: typeof import('./../src/components/base/collapse-links/VCollapseLinks.vue')['default']
-    VCollapseModel: typeof import('./../src/components/base/accordion/VCollapseModel.vue')['default']
-    VControl: typeof import('./../src/components/base/form/VControl.vue')['default']
-    VCreditCard: typeof import('./../src/components/base/plugins/VCreditCard.vue')['default']
-    VDropdown: typeof import('./../src/components/base/dropdown/VDropdown.vue')['default']
-    VField: typeof import('./../src/components/base/form/VField.vue')['default']
-    VFilePond: typeof import('./../src/components/base/plugins/VFilePond.vue')['default']
-    VFlex: typeof import('./../src/components/base/flex/VFlex.vue')['default']
-    VFlexItem: typeof import('./../src/components/base/flex/VFlexItem.vue')['default']
-    VFlexPagination: typeof import('./../src/components/base/pagination/VFlexPagination.vue')['default']
-    VFlexTable: typeof import('./../src/components/base/table/VFlexTable.vue')['default']
-    VFlexTableCell: typeof import('./../src/components/base/table/VFlexTableCell.vue')['default']
-    VFlexTableSortColumn: typeof import('./../src/components/base/table/VFlexTableSortColumn.vue')['default']
-    VFlexTableToolbar: typeof import('./../src/components/base/table/VFlexTableToolbar.vue')['default']
-    VFlexTableWrapper: typeof import('./../src/components/base/table/VFlexTableWrapper.vue')['default']
-    VGrid: typeof import('./../src/components/base/grid/VGrid.vue')['default']
-    VGridItem: typeof import('./../src/components/base/grid/VGridItem.vue')['default']
-    VIcon: typeof import('./../src/components/base/icon/VIcon.vue')['default']
-    VIconBox: typeof import('./../src/components/base/icon/VIconBox.vue')['default']
-    VIconButton: typeof import('./../src/components/base/button/VIconButton.vue')['default']
-    VIconWrap: typeof import('./../src/components/base/icon/VIconWrap.vue')['default']
-    VIMaskInput: typeof import('./../src/components/base/plugins/VIMaskInput.vue')['default']
-    VInput: typeof import('./../src/components/base/form/VInput.vue')['default']
-    VLabel: typeof import('./../src/components/base/form/VLabel.vue')['default']
-    VLoader: typeof import('./../src/components/base/loader/VLoader.vue')['default']
-    VMarkdownEditor: typeof import('./../src/components/base/markdown/VMarkdownEditor.vue')['default']
-    VMarkdownPreview: typeof import('./../src/components/base/markdown/VMarkdownPreview.vue')['default']
-    VMessage: typeof import('./../src/components/base/message/VMessage.vue')['default']
-    VModal: typeof import('./../src/components/base/modal/VModal.vue')['default']
-    VOptgroup: typeof import('./../src/components/base/form/VOptgroup.vue')['default']
-    VOption: typeof import('./../src/components/base/form/VOption.vue')['default']
-    VPageContent: typeof import('./../src/components/base/layouts/VPageContent.vue')['default']
-    VPageContentWrapper: typeof import('./../src/components/base/layouts/VPageContentWrapper.vue')['default']
-    VPeity: typeof import('./../src/components/base/plugins/VPeity.vue')['default']
-    VPhotosSwipe: typeof import('./../src/components/base/plugins/VPhotosSwipe.vue')['default']
-    VPlaceholderPage: typeof import('./../src/components/base/placeholder/VPlaceholderPage.vue')['default']
-    VPlaceholderSection: typeof import('./../src/components/base/placeholder/VPlaceholderSection.vue')['default']
-    VPlaceload: typeof import('./../src/components/base/loader/VPlaceload.vue')['default']
-    VPlaceloadAvatar: typeof import('./../src/components/base/loader/VPlaceloadAvatar.vue')['default']
-    VPlaceloadText: typeof import('./../src/components/base/loader/VPlaceloadText.vue')['default']
-    VPlaceloadWrap: typeof import('./../src/components/base/loader/VPlaceloadWrap.vue')['default']
-    VPlyr: typeof import('./../src/components/base/plugins/VPlyr.vue')['default']
-    VProgress: typeof import('./../src/components/base/progress/VProgress.vue')['default']
-    VRadio: typeof import('./../src/components/base/form/VRadio.vue')['default']
-    VRangeRating: typeof import('./../src/components/base/form/VRangeRating.vue')['default']
-    VReloadPrompt: typeof import('./../src/components/base/modal/VReloadPrompt.vue')['default']
-    VSelect: typeof import('./../src/components/base/form/VSelect.vue')['default']
-    VSnack: typeof import('./../src/components/base/snack/V-Snack.vue')['default']
-    VSwitchBlock: typeof import('./../src/components/base/form/VSwitchBlock.vue')['default']
-    VSwitchSegment: typeof import('./../src/components/base/form/VSwitchSegment.vue')['default']
-    VTabs: typeof import('./../src/components/base/tabs/VTabs.vue')['default']
-    VTag: typeof import('./../src/components/base/tags/VTag.vue')['default']
-    VTags: typeof import('./../src/components/base/tags/VTags.vue')['default']
-    VTextarea: typeof import('./../src/components/base/form/VTextarea.vue')['default']
-    VTextEllipsis: typeof import('./../src/components/base/typography/VTextEllipsis.vue')['default']
-    VViewWrapper: typeof import('./../src/components/base/layouts/VViewWrapper.vue')['default']
-    VVivus: typeof import('./../src/components/base/plugins/VVivus.vue')['default']
+    VAccordion: typeof import('./../src/components/base/VAccordion.vue')['default']
+    VAccordionImage: typeof import('./../src/components/base/VAccordionImage.vue')['default']
+    VAction: typeof import('./../src/components/base/VAction.vue')['default']
+    VAnimatedCheckbox: typeof import('./../src/components/base/VAnimatedCheckbox.vue')['default']
+    VAvatar: typeof import('./../src/components/base/VAvatar.vue')['default']
+    VAvatarStack: typeof import('./../src/components/base/VAvatarStack.vue')['default']
+    VBillboardJS: typeof import('./../src/components/base/VBillboardJS.vue')['default']
+    VBlock: typeof import('./../src/components/base/VBlock.vue')['default']
+    VBreadcrumb: typeof import('./../src/components/base/VBreadcrumb.vue')['default']
+    VButton: typeof import('./../src/components/base/VButton.vue')['default']
+    VButtons: typeof import('./../src/components/base/VButtons.vue')['default']
+    VCard: typeof import('./../src/components/base/VCard.vue')['default']
+    VCardAction: typeof import('./../src/components/base/VCardAction.vue')['default']
+    VCardAdvanced: typeof import('./../src/components/base/VCardAdvanced.vue')['default']
+    VCardMedia: typeof import('./../src/components/base/VCardMedia.vue')['default']
+    VCardSocial: typeof import('./../src/components/base/VCardSocial.vue')['default']
+    VCheckbox: typeof import('./../src/components/base/VCheckbox.vue')['default']
+    VCollapse: typeof import('./../src/components/base/VCollapse.vue')['default']
+    VCollapseLinks: typeof import('./../src/components/base/VCollapseLinks.vue')['default']
+    VCollapseModel: typeof import('./../src/components/base/VCollapseModel.vue')['default']
+    VControl: typeof import('./../src/components/base/VControl.vue')['default']
+    VCreditCard: typeof import('./../src/components/base/VCreditCard.vue')['default']
+    VDataTable: typeof import('./../src/components/base/VDataTable.vue')['default']
+    VDropdown: typeof import('./../src/components/base/VDropdown.vue')['default']
+    VField: typeof import('./../src/components/base/VField.vue')['default']
+    VFilePond: typeof import('./../src/components/base/VFilePond.vue')['default']
+    VFlex: typeof import('./../src/components/base/VFlex.vue')['default']
+    VFlexItem: typeof import('./../src/components/base/VFlexItem.vue')['default']
+    VFlexPagination: typeof import('./../src/components/base/VFlexPagination.vue')['default']
+    VFlexTable: typeof import('./../src/components/base/VFlexTable.vue')['default']
+    VFlexTableCell: typeof import('./../src/components/base/VFlexTableCell.vue')['default']
+    VFlexTableSortColumn: typeof import('./../src/components/base/VFlexTableSortColumn.vue')['default']
+    VFlexTableToolbar: typeof import('./../src/components/base/VFlexTableToolbar.vue')['default']
+    VFlexTableWrapper: typeof import('./../src/components/base/VFlexTableWrapper.vue')['default']
+    VGrid: typeof import('./../src/components/base/VGrid.vue')['default']
+    VGridItem: typeof import('./../src/components/base/VGridItem.vue')['default']
+    VIcon: typeof import('./../src/components/base/VIcon.vue')['default']
+    VIconBox: typeof import('./../src/components/base/VIconBox.vue')['default']
+    VIconButton: typeof import('./../src/components/base/VIconButton.vue')['default']
+    VIconWrap: typeof import('./../src/components/base/VIconWrap.vue')['default']
+    VIMaskInput: typeof import('./../src/components/base/VIMaskInput.vue')['default']
+    VInput: typeof import('./../src/components/base/VInput.vue')['default']
+    VLabel: typeof import('./../src/components/base/VLabel.vue')['default']
+    VLoader: typeof import('./../src/components/base/VLoader.vue')['default']
+    VMarkdownEditor: typeof import('./../src/components/base/VMarkdownEditor.vue')['default']
+    VMarkdownPreview: typeof import('./../src/components/base/VMarkdownPreview.vue')['default']
+    VMessage: typeof import('./../src/components/base/VMessage.vue')['default']
+    VModal: typeof import('./../src/components/base/VModal.vue')['default']
+    VMultiselect: typeof import('./../src/components/base/VMultiselect.vue')['default']
+    VOptgroup: typeof import('./../src/components/base/VOptgroup.vue')['default']
+    VOption: typeof import('./../src/components/base/VOption.vue')['default']
+    VPageContent: typeof import('./../src/components/base/VPageContent.vue')['default']
+    VPageContentWrapper: typeof import('./../src/components/base/VPageContentWrapper.vue')['default']
+    VPeity: typeof import('./../src/components/base/VPeity.vue')['default']
+    VPhotosSwipe: typeof import('./../src/components/base/VPhotosSwipe.vue')['default']
+    VPlaceholderPage: typeof import('./../src/components/base/VPlaceholderPage.vue')['default']
+    VPlaceholderSection: typeof import('./../src/components/base/VPlaceholderSection.vue')['default']
+    VPlaceload: typeof import('./../src/components/base/VPlaceload.vue')['default']
+    VPlaceloadAvatar: typeof import('./../src/components/base/VPlaceloadAvatar.vue')['default']
+    VPlaceloadText: typeof import('./../src/components/base/VPlaceloadText.vue')['default']
+    VPlaceloadWrap: typeof import('./../src/components/base/VPlaceloadWrap.vue')['default']
+    VPlyr: typeof import('./../src/components/base/VPlyr.vue')['default']
+    VProgress: typeof import('./../src/components/base/VProgress.vue')['default']
+    VRadio: typeof import('./../src/components/base/VRadio.vue')['default']
+    VRangeRating: typeof import('./../src/components/base/VRangeRating.vue')['default']
+    VReloadPrompt: typeof import('./../src/components/base/VReloadPrompt.vue')['default']
+    VSelect: typeof import('./../src/components/base/VSelect.vue')['default']
+    VSnack: typeof import('./../src/components/base/VSnack.vue')['default']
+    VSwitchBlock: typeof import('./../src/components/base/VSwitchBlock.vue')['default']
+    VSwitchSegment: typeof import('./../src/components/base/VSwitchSegment.vue')['default']
+    VTabs: typeof import('./../src/components/base/VTabs.vue')['default']
+    VTag: typeof import('./../src/components/base/VTag.vue')['default']
+    VTags: typeof import('./../src/components/base/VTags.vue')['default']
+    VTextarea: typeof import('./../src/components/base/VTextarea.vue')['default']
+    VTextEllipsis: typeof import('./../src/components/base/VTextEllipsis.vue')['default']
+    VViewWrapper: typeof import('./../src/components/base/VViewWrapper.vue')['default']
+    VVivus: typeof import('./../src/components/base/VVivus.vue')['default']
   }
 }
